Fix tf leak on marker(array) removal, SceneNode
MarkerArray client was not calling the removeTF() cleanup func in the
other marker-removal actions/callbacks (besides DELETE ALL).

Further, the removeTF() SceneNode cleanup function was passing the
wrong func to tfClient's unsubscribe, because the subscribe callback
was created with an (unreferencable) anonymous func. Had to reintroduce
the this.tfCallback in order to be able to properly refer to the
callback AND scope the unsubscribe to the right 'this'.